+++ src/types/action-trace.ts
@@ -1,7 +1,6 @@
-import { DbOp } from "./table-delta"
-import { RAMOp, DTrxOp, TableOp } from "./transaction"
+import { RamOp, DTrxOp, TableOp, Int64, Uint64 } from "./common"
 
-export type ActionTraceData<T = Record<string, any>, D = Record<string, any>> = {
+export type ActionTraceData<T = Record<string, any>> = {
   block_num: number
   block_id: string
   block_time: string
@@ -9,8 +8,8 @@ export type ActionTraceData<T = Record<string, any>, D = Record<string, any>> = 
   idx: number
   depth: number
   trace: ActionTrace<T>
-  dbops?: DbOp<D>[]
-  ramops?: RAMOp[]
+  dbops?: ActionTraceDbOp[]
+  ramops?: RamOp[]
   dtrxops?: DTrxOp[]
   tableops?: TableOp[]
 }
@@ -59,21 +58,67 @@ export type AccountRamDelta = {
 }
 
 /**
- * A `int64_t` natively in `nodeos` but can become a string when > 32 bits number
- * due to how `nodeos` serialize number to JSON.
+ * The `dbops` array out of an [[ActionTraceData]] message is completely
+ * different than other [[DbOp]] found in dfuse API (like on [[TableDeltaData]]
+ * or [[TransactionLifecycle]]).
  *
- * This is like because JavaScript largest number possible is 53 bits large which
- * make it impossible to hold a full `int64_t` type. To overcome that, `nodeos`
- * will output a string when number is too large to preserve precision.
+ * One for `opayer` or `npayer` will always be present depending on the
+ * operation, same thing for the `old` and `new` fields:
+ *
+ * - When `op == "INS"`, `npayer` and `new` are present
+ * - When `op == "UPD"`, `opayer`, `old`, `npayer` and `new` are present
+ * - When `op == "REM"`, `opayer` and `old` are present
+ *
+ * The `old` and `new` fields are the hexadecimal string encoded
+ * representing the row in binary format.
+ *
+ * Check the [Decode Hex Data using eosjs example](https://github.com/dfuse-io/example-eosjs-decode-hex)
+ * for a way to transform the hexadecimal string into a JSON
+ * structure representing the row.
+ *
+ * @see https://github.com/dfuse-io/example-eosjs-decode-hex
  */
-export type Int64 = number | string
+export type ActionTraceDbOp = {
+  op: "INS" | "UPD" | "REM"
+  action_idx: number
 
-/**
- * A `uint64_t` natively in `nodeos` but can become a string when > 32 bits number
- * due to how `nodeos` serialize number to JSON.
- *
- * This is like because JavaScript largest number possible is 53 bits large which
- * make it impossible to hold a full `uint64_t` type. To overcome that, `nodeos`
- * will output a string when number is too large to preserve precision.
- */
-export type Uint64 = number | string
+  /**
+   * The account which was the old payer of the row. Present when `op == "UPD" | "REM"`.
+   */
+  opayer?: string
+
+  /**
+   * The account which is the new payer of the row. Present when `op == "UPD" | "REM"`.
+   */
+  npayer?: string
+
+  /**
+   * The full path of the database row, it's a string with four elements
+   * of the row path separated with the `/` character being respectively
+   * from left to right: account, scope, table, row primary key (name encoded).
+   *
+   * ```
+   * "eosio.token/trustdicewin/accounts/........ehbo5"
+   *
+   * {
+   *   account: "eosio.token",
+   *   scope: "trustdicewin",
+   *   table: "accounts",
+   *   key: "........ehbo5", // Name encoded value representing `EOS` SymbolCode
+   * }
+   * ```
+   */
+  path: string
+
+  /**
+   * The hexadecimal string encoded representing the old version of the
+   * row in binary format. Present when `op == "UPD" | "REM"`.
+   */
+  old?: string
+
+  /**
+   * The hexadecimal string encoded representing the new version of the
+   * row in binary format. Present when `op == "INS" | "UPD"`.
+   */
+  new?: string
+}